@@ -1,6 +1,6 @@
 import { Participation } from './Participation';
 
-interface Olympic {
+export interface Olympic {
     id: number;
     country: string;
     participations: Participation[];
