@@ -1,4 +1,4 @@
-interface Participation {
+export interface Participation {
   id: number;
   year: number;
   city: string;
